Move gc_disabled decorator into alias module
